pyproject.toml
[project]
name = "hdbase"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "celery==5.4.0",
    "channels==4.2.2",
    "channels-redis==4.2.1",
    "daphne==4.1.2",
    "django==5.1.7",
    "django-bootstrap-datepicker-plus>=5.0.5",
    "django-filter>=25.1",
    "django-froala-editor>=4.5.2",
    "django-jsonform>=2.23.2",
    "django-redis>=5.4.0",
    "django-select2>=8.4.1",
    "jinja2==3.1.4",
    "pillow==11.1.0",
    "psutil>=7.0.0",
    "pyorthanc>=1.21.0",
    "redis==5.2.1",
]
